llm/requirements.txt
pika
langchain-core
langchain
requests
python-dotenv